switched fetch request from main container to login container, still doesn't redirect
