+++ client/containers/LoginContainer.jsx
@@ -1,8 +1,13 @@
 import React, { Component, useState, useEffect } from 'react';
 import { connect } from 'react-redux';
-import { NavLink } from 'react-router-dom';
+import { NavLink, Redirect } from 'react-router-dom';
 import { login } from '../actions/actions';
 import { useDispatch } from 'react-redux';
+// import { Router } from 'express';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import MainContainer from './MainContainer.jsx';
+
+
 
 
 // redirect tag 
@@ -11,13 +16,13 @@ const LoginContainer = () => {
   // hook in place of mapDispatchToProps
   const dispatch = useDispatch();
   // useState hook to hold form data
-  const [isFetch, setIsFetch] = useState(false);
+  // const [isFetch, setIsFetch] = useState(false);
   const [fields, setFields] = useState({
     username: '',
     password: '',
   });
 
-  const handleClick = (e) => {
+  const handleChange = (e) => {
     const targetName = e.target.name;
     const value = e.target.value;
     setFields({
@@ -28,8 +33,44 @@ const LoginContainer = () => {
 
   // **************need real data, need to uncomment line 41 and delete 44-49 *************
   const fetchLogin = () => {
-    setIsFetch(true)
+    console.log('fetchlogin clicked!')
+    const reqData = { 
+      username: fields.username,
+      password: fields.password
+    }
+    fetch('/api/', {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify(reqData)})
+      .then((res) => {
+        return res.json();
+      })
+      .then((data) => {
+        console.log('data returned from fetch', data)
+        const payload = {};
+        payload.user = data.user;
+        payload.projects = data.projects;
+        payload.categories = data.categories;
+        payload.timerActivity = data.timerHistory;
+
+        dispatch(login(payload));
+        console.log('dispatch finished')
+      })
+      .then(() => (
+        <Router>
+          <Route exact path="/user" component={MainContainer}/>
+        </Router>
+      ))
+      .catch(err => console.log('there was an error fetching data: ', err))
+
+    // send fetch request, post, sending usernamse and pw
+    // get response and dispatch data to set state
+    // then redirect to /user
+
   }
+  // setIsFetch(true)
   // useEffect(() => {
   //   // object to send in fetch request
   //   const reqData = { 
@@ -47,66 +88,6 @@ const LoginContainer = () => {
   //       // return res.json();
   //     // })
   //     // .then((data) => {
-  //       // dummy data for test
-  //         const data = {
-  //           user: {
-  //             _id: 1,
-  //             account_name: 'tom',
-  //         },
-  //           projects: [{
-  //             _id: 1,
-  //             title: 'Not specified',
-  //           }, {           
-  //             _id: 2,
-  //             title: 'Project 1',
-  //           }],
-  //           categories: [{
-  //             _id: 1,
-  //             title: 'Coding',
-  //           }, {
-  //             _id: 2,
-  //             title: 'Debug',
-  //           }, {
-  //             _id: 3,
-  //             title: 'Meetings',
-  //           }, {
-  //             _id: 4,
-  //             title: 'QA',
-  //           }, {
-  //             _id: 5,
-  //             title: 'Code Review',
-  //           }, {
-  //             _id: 6,
-  //             title: 'Reasearch',
-  //           }, {
-  //             _id: 7,
-  //             title: 'Write Documentation',
-  //           }, {
-  //             _id: 7,
-  //             title: 'Other',
-  //           }],
-  //           timerHistory: [{
-  //             time_spent: 5000,
-  //             updated_at: new Date().toString(),
-  //             category_id: 3,
-  //             project_id: 2,
-  //           }, {
-  //             time_spent: 4000,
-  //             updated_at: new Date().toString(),
-  //             category_id: 1,
-  //             project_id: 2,
-  //           }, {
-  //             time_spent: 2000,
-  //             updated_at: new Date().toString(),
-  //             category_id: 3,
-  //             project_id: 1,
-  //           }, {
-  //             time_spent: 4000,
-  //             updated_at: new Date().toString(),
-  //             category_id: 2,
-  //             project_id: 2,
-  //           }]
-  //         }
   //         const payload = {};
   //         payload.user = data.user;
   //         payload.projects = data.projects;
@@ -143,7 +124,7 @@ const LoginContainer = () => {
             placeholder="Username"
             className="info"
             value={fields.username}
-            onChange={handleClick}
+            onChange={handleChange}
           />
           <input 
             id="password" 
@@ -151,14 +132,14 @@ const LoginContainer = () => {
             type="password"
             placeholder="Password"
             value={fields.password}
-            onChange={handleClick} 
+            onChange={handleChange} 
           />
           {/* when back end is working an rediresting to user page this nav link need to be deleted */}
-          <NavLink to="/user">
+          {/* <NavLink to="/user"> */}
             <button type="button" id="log-in" className="button" onClick={fetchLogin}>
               Log In
             </button>
-          </NavLink>
+          {/* </NavLink> */}
           <p>Login with <img src="https://github.githubassets.com/images/modules/logos_page/GitHub-Logo.png" height="20px"/> <img src="https://image.flaticon.com/icons/svg/25/25231.svg" width="25px" height="25px"/></p>
         </div>
         <div id="no-log-in">
